+++ Editor/HapticSourceEditor.cs
@@ -8,13 +8,17 @@ using UnityEngine;
 using Interhaptics;
 using Interhaptics.Internal;
 
-[CustomEditor(typeof(HapticSource))]
-public class HapticSourceEditor : Editor
+namespace Interhaptics.Editor
 {
-    SerializedProperty hapticMaterial;
-    SerializedProperty vibrationOffset;
-    SerializedProperty textureOffset;
-    SerializedProperty stiffnessOffset;
+
+[CustomEditor(typeof(HapticSource))]
+public class HapticSourceEditor : UnityEditor.Editor
+{
+    private SerializedProperty hapticMaterial;
+    private SerializedProperty vibrationOffset;
+    private SerializedProperty textureOffset;
+    private SerializedProperty stiffnessOffset;
+    private SerializedProperty debugMode;
 
     private void OnEnable()
     {
@@ -22,6 +26,7 @@ public class HapticSourceEditor : Editor
         vibrationOffset = serializedObject.FindProperty("vibrationOffset");
         textureOffset = serializedObject.FindProperty("textureOffset");
         stiffnessOffset = serializedObject.FindProperty("stiffnessOffset");
+        debugMode = serializedObject.FindProperty("debugMode");
     }
 
     public override void OnInspectorGUI()
@@ -40,6 +45,11 @@ public class HapticSourceEditor : Editor
         GUIContent stiffnessOffsetLabel = new GUIContent("Stiffness offset", "The offset applied to the stiffness of the haptic effect.");
         EditorGUILayout.PropertyField(stiffnessOffset, stiffnessOffsetLabel);
 
+        GUIContent debugModeLabel = new GUIContent("Debug mode", "Shows debug messages in Editor if checked.");
+        EditorGUILayout.PropertyField(debugMode, debugModeLabel);
+
         serializedObject.ApplyModifiedProperties();
     }
+}
+
 }